Main

Description

Documentation

main :: IO ()

The main function is just putting together everything else in the fuckin file tbh

binary :: String -> Integer

binary takes a string representing a base-2 number and returns its value as a base-10 Integer it more or less just takes a string, reverses it, and applies binary' to that with an initial value of 0

binary' :: Integer -> String -> Integer

binary' does the heavy lifting here; if it's at the end of the string

url :: [String]

url is just the test instance for the above