Main

Description

Documentation

main :: IO ()

What main does is set up the window, create the "global variables" angle, delta, and pos, and sets up the various callback functions, resulting in a call to mainLoop